// src/models/page-result-store-resp.ts
// tslint:disable
/**
 * geo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { StoreResp } from './store-resp';

/**
 * 
 * @export
 * @interface PageResultStoreResp
 */
export interface PageResultStoreResp {
    /**
     * 
     * @type {Array<StoreResp>}
     * @memberof PageResultStoreResp
     */
    list?: Array<StoreResp>;
    /**
     * 
     * @type {number}
     * @memberof PageResultStoreResp
     */
    total?: number;
    /**
     * 
     * @type {number}
     * @memberof PageResultStoreResp
     */
    page?: number;
    /**
     * 
     * @type {number}
     * @memberof PageResultStoreResp
     */
    size?: number;
    /**
     * 
     * @type {number}
     * @memberof PageResultStoreResp
     */
    pages?: number;
}