// src/models/staff-update-req.ts
// tslint:disable
/**
 * geo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 员工更新请求
 * @export
 * @interface StaffUpdateReq
 */
export interface StaffUpdateReq {
    /**
     * 手机号
     * @type {string}
     * @memberof StaffUpdateReq
     */
    phone?: string;
    /**
     * 微信号
     * @type {string}
     * @memberof StaffUpdateReq
     */
    wechat?: string;
    /**
     * 新密码
     * @type {string}
     * @memberof StaffUpdateReq
     */
    password?: string;
    /**
     * 姓名
     * @type {string}
     * @memberof StaffUpdateReq
     */
    name?: string;
    /**
     * 角色
     * @type {string}
     * @memberof StaffUpdateReq
     */
    role?: StaffUpdateReqRoleEnum;
    /**
     * 所属门店ID
     * @type {number}
     * @memberof StaffUpdateReq
     */
    storeId?: number;
    /**
     * 队长ID
     * @type {number}
     * @memberof StaffUpdateReq
     */
    leaderId?: number;
    /**
     * 状态
     * @type {number}
     * @memberof StaffUpdateReq
     */
    status?: number;
}

/**
    * @export
    * @enum {string}
    */
export enum StaffUpdateReqRoleEnum {
    ADMIN = 'ADMIN',
    TEAMLEADER = 'TEAM_LEADER',
    SALESMAN = 'SALESMAN'
}