// src/models/page-result-staff-resp.ts
// tslint:disable
/**
 * geo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { StaffResp } from './staff-resp';

/**
 * 
 * @export
 * @interface PageResultStaffResp
 */
export interface PageResultStaffResp {
    /**
     * 
     * @type {Array<StaffResp>}
     * @memberof PageResultStaffResp
     */
    list?: Array<StaffResp>;
    /**
     * 
     * @type {number}
     * @memberof PageResultStaffResp
     */
    total?: number;
    /**
     * 
     * @type {number}
     * @memberof PageResultStaffResp
     */
    page?: number;
    /**
     * 
     * @type {number}
     * @memberof PageResultStaffResp
     */
    size?: number;
    /**
     * 
     * @type {number}
     * @memberof PageResultStaffResp
     */
    pages?: number;
}